requirements.txt
requests
beautifulsoup4
python-dotenv
Flask
celery
redis